Move ndarray to separate submodules
