+++ bindings/pyroot/ROOT.py
@@ -22,11 +22,6 @@ __author__  = 'Wim Lavrijsen (WLavrijsen@lbl.gov)'
 ### system and interpreter setup ------------------------------------------------
 import os, sys, types
 import cppyy
-# Import numpy lazily (needed for `ndarray` class)
-try:
-    import numpy
-except:
-    raise ImportError("Failed to import numpy.")
 
 ## there's no version_info in 1.5.2
 if sys.version[0:3] < '2.2':
@@ -374,23 +369,6 @@ def _TTreeAsMatrix(self, columns=None, exclude=None, dtype="double", return_labe
         return reshaped_matrix_np
 
 
-class ndarray(numpy.ndarray):
-    """
-    A wrapper class that inherits from numpy.ndarray and allows to attach the
-    result pointer of the `Take` action in an `RDataFrame` event loop to the
-    collection of values returned by that action.
-    """
-    def __new__(cls, numpy_array, result_ptr):
-        obj = numpy.asarray(numpy_array).view(cls)
-        obj.result_ptr = result_ptr
-        obj.__class__.__name__ = "numpy.array"
-        return obj
-
-    def __array_finalize__(self, obj):
-        if obj is None: return
-        self.result_ptr = getattr(obj, "result_ptr", None)
-
-
 # RDataFrame.AsNumpy feature
 def _RDataFrameAsNumpy(df, columns=None, exclude=None):
     """Read-out the RDataFrame as a collection of numpy arrays.
@@ -416,6 +394,13 @@ def _RDataFrameAsNumpy(df, columns=None, exclude=None):
     Returns:
         dict: Dict with column names as keys and 1D numpy arrays with content as values
     """
+    # Import numpy and numpy.array derived class lazily
+    try:
+        import numpy
+        from _helpers import ndarray
+    except:
+        raise ImportError("Failed to import numpy during call of RDataFrame.AsNumpy.")
+
     # Find all column names in the dataframe if no column are specified
     if not columns:
         columns = [c for c in df.GetColumnNames()]
